+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add stale while revalidate and stale if error cache header
+ * Allow dynamic session "ttl" when using a remote storage
 
 6.0
 ---